Fitting preparation (#8)
* minimizers to sub dir

* adjustments to new Fitting folder

* code clarifications

* splitting up fitting tests

* cleaning

* cleaning

* cleaning

* cleaning

* from engine_ to minimizer_

* adjustments to small and large caps
