+++ examples_old/example1.py
@@ -3,7 +3,7 @@ __version__ = '0.1.0'
 
 import numpy as np
 
-from easyscience.Fitting.Fitting import Fitter
+from easyscience.Fitting import Fitter
 from easyscience.Objects.ObjectClasses import BaseObj
 from easyscience.Objects.ObjectClasses import Parameter
