+++ examples_old/example6_broken.py
@@ -9,7 +9,7 @@ from typing import Callable
 import numpy as np
 
 from easyscience import borg
-from easyscience.Fitting.Fitting import Fitter
+from easyscience.Fitting import Fitter
 from easyscience.Objects.Base import BaseObj
 from easyscience.Objects.Base import Parameter
 from easyscience.Objects.core import ComponentSerializer
