+++ examples_old/example_dataset2.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 
 import numpy as np
 from easyscience.Datasets.xarray import xr
-from easyscience.Fitting.Fitting import Fitter
+from easyscience.Fitting import Fitter
 from easyscience.Objects.ObjectClasses import BaseObj
 from easyscience.Objects.ObjectClasses import Parameter
